created separate function fn_glibcfix
First function to be migrated to its own file. This will slowly be done
over time to more functions as changes are made. The idea is to reduce
the requirement to copy the same changes to every script many times.

This fixes the following issue: libstdc++.so.6 fix
https://github.com/dgibbs64/linuxgameservers/issues/203

This is the first release of this new file function feature any bugs
please report. To apply the fix to existing servers you will need to run
the installer again.