Migrated to new split functions system
All scripts now have the new split functions system in-place.

Some functions now have there own file rather than being in the main
script file.
If the main script requires the file is will automatically download it
in the the functions dir for use.

Only a few functions currently do this. However if a change is made to a
function it will be migrated. In the long term this will reduce the need
for me to copy the same code over 20 times and drastically reduce the
time it takes me to make changes.