Major rename of functions
* All functions names have been reviewed and renamed to make it easier
to identify what the function is for. This is for my benefit as there
are now over 50 functions to manage.

* New details feature. It has been reworked to add more useful info and
fix any bugs with details not always being displayed.

* Main script file e.g csgoserver should no longer require any updates
from this point. All code modifications should be made from the
functions. Any future edits would relate to server specific config.

* servername variable has been moved to its own function if name is
gathered from a config file. The same applies for ports, usernames and
passwords.

* new function fn_details_distro gathers all the required OS info for
fn_details and other functions that require this info

* Various improvements and bug fixes that I cannot remember.