Update fn_stop
Modified stop function to allow for graceful shutdown of a 7 Days To Die server.  The file will still work for other game servers as it checks for the game name to be "7 Days To Die" as set in the sdtdserver script file and it also uses the new telnet variables I added to the sdtdserver script.
If not being run for 7 Days To Die, it reverts to the previous stop method of killing the server process.