Update fn_validate
Made a custom change for 7 Days To Die.  If user wants or has Alloc's server fixes installed, they need to set the variable in the sdtdserver script to true and then the added coded in fn_validate will now download and install Alloc's fixes automatically.
Could possibly be put in a separate file and called by a function, but validate isn't run constantly so the extra few lines of code shouldn't matter.