Changed tar exclude to be the configured dir
Updated the tar exclude to be the configured backup dir, instead of just a hardcoded "backup". This allows for people to change the backup directory location to still be in the game server dir, but in a custom directory. For me specifically it's an NFS mounted directory to backup to my NAS with the cloud backup then enabled on that dir.