Change logfile used for update check
This changes the logfile that is used from the log generated by tmux to the log generated by the factorio binary. As logging is disabled in tmux 1.8, those users which may be stuck on that version cannot use the update functionality, as there is no log to go off of.