Update _default.cfg
Replacing the steam login details with an anonymous login will allow the install of the 7dtd server to complete successfully without having to edit common.cfg beforehand (which is only created when you run ./sdtdserver install which fails). The webpage information is misleading currently, telling you to edit a file that doesn't yet exist.