Fixed tmux sessions killed when they should not
When graceful works, probably because of pipe-pane or for whatever reason, the tmux session dies. But fn_stop_tmux was ran anyways, which then caused a command like
tmux kill-session -t csgoserver 
to affect the closest matching alternate, like csgoserver-2 tmux session.
With this change, we make sure to run the session kill only when required.