Update to allow use of backup exclusion file for ommiting content from backups
I've reworked the backup command to use a flat file that is used for file and folder exclusions in an attempt to allow admins to exclude files or folders from their backups. The idea being they can exclude game files that will be replaced by repairing the game server and it helps keep their backups smaller.

I've also added the necessary exclude file variable and a check for the "backupexclude.cfg" in the same fashion as "common.cfg" and "instance.cfg" so it should fold into current installations without issue.