feat(rw): migrate rising world to unity engine (#4261)
* refactor: update default executable name for Rising World server

The default executable name for the Rising World server has been updated from "server.jar" to "RisingWorldServer.x64". This change ensures that the correct executable is used when launching the game server.

* refactor(rw): no longer java

The code in `info_messages.sh` has been refactored to simplify the port command for multiple game servers. This change improves readability and maintainability.

* refactor: simplify info_game_rw function

The info_game_rw function has been refactored to remove redundant code and improve readability. The configuration values are now retrieved using the fn_info_game_keyvalue_pairs function, which reduces duplication and improves maintainability. Additionally, the default values for configip, gamemode, maxplayers, port, queryport, rconport, seed, servername, and worldname have been set to appropriate fallback values if they are not found in the server configuration file.

* refactor: remove unnecessary game ports and query HTTP port

The code changes in this commit refactor the `info_messages.sh` file by removing unnecessary game ports and the query HTTP port. This simplifies the code and improves readability.

* refactor: update queryport calculation in info_game.sh

The code change updates the calculation of the queryport variable in the info_game.sh script. Instead of assigning it the same value as port, it now subtracts 1 from the port value. This ensures that queryport is set correctly for further use in the script.

* feat: add branch configuration option for rwserver

This commit adds a new configuration option `branch` to the `_default.cfg` file in the `rwserver` directory. The default value is set to "unity". This allows users to specify a specific branch for SteamCMD when installing or updating the server.

* refactor: update default game server configuration

The default game server configuration file has been updated to reflect the following changes:
- Removed the predefined parameter `javaram`
- Updated the value of `consoleinteract` to "yes"
- Updated the values of `engine` and `glibc`

These changes ensure that the default configuration is more up-to-date and aligned with the desired behavior for a Rising World game server.

* refactor: remove unnecessary code for game+1, game+2, and game+3

The code changes in this commit involve removing unnecessary code related to the variables port2, port3, and port4. These variables were used to display information about additional games (game+1, game+2, and game+3) in the output. However, since these variables are not being used elsewhere in the script, they have been removed along with the corresponding echo statements. This refactor simplifies the code by eliminating redundant logic.