feat: various UI changes (#4326)
* refactor: remove unnecessary comments in default server configurations

The commit removes unnecessary comments from the default server configurations. The comments were providing instructions to edit the configuration files with care, but they are not needed and can be safely removed.

* refactor: update server names in default configuration files

- Updated the servername value in multiple default configuration files to "LinuxGSM" for consistency.
- Refactored the code to reflect the changes made.

* refactor: rename fn_script_log_fatal to fn_script_log_faail

to create more consistency with naming

* feat: improve console output formatting

- Added separator function to improve readability of console output.
- Replaced multiple instances of manual separators with the new function.
- Updated commit message to follow conventional commit standard.

* refactor: remove unnecessary fn_sleep_time calls

This commit removes the fn_sleep_time function calls that are no longer needed in multiple modules. The fn_sleep_time function was used to introduce a delay, but it is no longer necessary for the current implementation.

* better standardise random

* removed bold where not required