refactor: improve steamcmd dependency check
The code in `check_deps.sh` has been refactored to improve the steamcmd dependency check. The condition for checking if the non-free repository is enabled in Debian has been updated to use a more accurate regular expression. This change ensures that the correct dependencies are required and improves the accuracy of the check.